chore: cli cleanup
